+++ client/__stories__/components/post/CommentInput.stories.tsx
@@ -1,21 +1,28 @@
 import React from 'react';
-import CommentList from '@/components/post/CommentList/CommentList';
+import CommentInput from '@/components/post/CommentList/CommentInput';
 import { ComponentStory } from '@storybook/react';
-import { COMMENT, COMMENT2 } from '@mocks/data/comments';
+import { rest } from 'msw';
+import { mockApiURL } from '@mocks/apis/utils';
 
 export default {
-  component: CommentList,
-  title: 'post/CommentList',
+  component: CommentInput,
+  title: 'post/CommentInput',
   parameters: {
     actions: { argTypesRegex: '^on.*' },
   },
 };
 
-const Template: ComponentStory<typeof CommentList> = (args) => (
-  <CommentList {...args} />
+const Template: ComponentStory<typeof CommentInput> = (args) => (
+  <CommentInput {...args} />
 );
 
 export const Default = Template.bind({});
-Default.args = {
-  comments: [COMMENT, COMMENT2],
+
+export const WithoutUser = Template.bind({});
+WithoutUser.parameters = {
+  msw: [
+    rest.get(mockApiURL('/users/profile'), (req, res, ctx) =>
+      res(ctx.status(403)),
+    ),
+  ],
 };
